Fixes mod name, changes from PatchOperationAdd to PatchOperationAddModExtension
Mod name was "Android Tiers", fixed to "Android tiers". Using PatchOperationAddModExtension instead of PatchOperationAdd.

A point could be made to merge all the patches into a single 'or' patch, as that would be faster, but I dont know if the speed bonus is desirable considering it'll make the patch uglier.